fix: clean up code
