+++ frontend/modules/dashboard/components/dashboard-edit-form.tsx
@@ -15,19 +15,13 @@ import { useGetSingleUser } from '../hooks/use-get-single-user';
 
 export const DashboardEditForm: React.FC<{ id: number }> = ({ id }) => {
   const { data } = useGetSingleUser(id);
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-    reset
-  } = useEditForm();
+  const { register, handleSubmit, formState, reset } = useEditForm();
+
   const [loading, setLoading] = useState(false);
   const { toast } = useToast();
   const router = useRouter();
 
-  useEffect(() => {
-    if (data) reset({ ...data, password: '' });
-  }, [data, reset]);
+  const { errors, isDirty } = formState;
 
   const onSubmit = async (formData: EditFormData) => {
     setLoading(true);
@@ -36,20 +30,30 @@ export const DashboardEditForm: React.FC<{ id: number }> = ({ id }) => {
       invalidateQuery();
       toast({ title: 'Success', variant: 'success' });
       router.push(ROUTES.DASHBOARD);
-    } catch (error) {
-      toast({ title: 'Something went wrong', variant: 'destructive' });
+    } catch (e) {
+      if (e instanceof Error) toast({ title: e.message, variant: 'destructive' });
       // eslint-disable-next-line no-console
-      console.error(error);
+      console.error(e);
     } finally {
       setLoading(false);
     }
   };
 
+  useEffect(() => {
+    if (data) reset({ ...data, password: '' });
+  }, [data, reset]);
+
   return (
     <form className="space-y-4 md:space-y-6" onSubmit={handleSubmit(onSubmit)}>
       <Input label="Username" {...register('username')} error={errors.username?.message} placeholder="Ex: John Doe" />
 
-      <Input label="Email" {...register('email')} error={errors.email?.message} placeholder="username@gmail.com" />
+      <Input
+        label="Email"
+        {...register('email')}
+        error={errors.email?.message}
+        placeholder="username@gmail.com"
+        disabled
+      />
 
       <PasswordInput
         label="Password"
@@ -72,7 +76,7 @@ export const DashboardEditForm: React.FC<{ id: number }> = ({ id }) => {
         placeholder="••••••••"
       />
 
-      <Button type="submit" className="w-full" disabled={loading}>
+      <Button type="submit" className="w-full" disabled={loading || !isDirty}>
         Submit
       </Button>
     </form>
